add docker

+++ init.sh
@@ -1,4 +1,8 @@
-ln -s ~/Desktop/llvm/ /tmp/llvm-project
+if [ -f /.dockerenv ]; then
+	:
+else
+	ln -s ~/Desktop/llvm/ /tmp/llvm-project
+fi
 export LLVM_VIRT=/tmp/llvm-project/virtualroot
 export PATH=$PWD/dfsan-plugin:$LLVM_VIRT/bin:$PATH
 export LD_LIBRARY_PATH=$LLVM_VIRT/lib:$LD_LIBRARY_PATH
